refactor : 비즈니스 로직, 서비스 로직 분리
App 클래스에 존재했던 비즈니스 로직, 서비스 로직을 분리하여 각 클래스가
하나의 역할만 하도록 리팩토링
